fixed bugs and maybe fixed lint
