+++ clients/android/app/src/main/res/layout/activity_main_screen.xml
@@ -25,7 +25,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/detail_container"
-            android:layout_width="400dp"
+            android:layout_width="590dp"
             android:layout_height="match_parent"
             android:layout_weight="1" />
