+++ libs/content/workspace-ffi/src/apple/mod.rs
@@ -2,4 +2,3 @@ pub mod common_ffi;
 pub mod ios_ffi;
 pub mod keyboard;
 pub mod macos_ffi;
-pub mod resp;
